fix: added support for more chains

+++ features/home/impl/src/main/kotlin/io/element/android/features/home/impl/HomeEvents.kt
@@ -39,7 +39,7 @@ sealed interface HomeEvents {
     sealed interface HomeWalletEvents: HomeEvents
     data class LoadMoreTokens(val currentTokens: List<ZeroWalletToken>): HomeWalletEvents
     data class LoadMoreTransactions(val currentTransactions: List<ZeroWalletTransaction>): HomeWalletEvents
-    data class ViewWalletTransaction(val transactionId: String, val chainId: Int? = null): HomeWalletEvents
+    data class ViewWalletTransaction(val transactionId: String, val chainId: Long? = null): HomeWalletEvents
     data object OnWalletTransactionViewed: HomeWalletEvents
     data object ToggleWalletBalance: HomeWalletEvents
     data object RefreshWalletBalance: HomeWalletEvents
