+++ zero/src/main/java/io/element/android/support/zero/data/repository/StakeRepository.kt
@@ -16,21 +16,21 @@ import io.element.android.support.zero.network.model.response.wallet.ApiTransact
 
 interface StakeRepository {
 
-    suspend fun getTotalStaked(poolAddress: String, chainId: Int): String
+    suspend fun getTotalStaked(poolAddress: String, chainId: Long): String
 
-    suspend fun getStakingConfig(poolAddress: String, chainId: Int): ApiStakingConfig
+    suspend fun getStakingConfig(poolAddress: String, chainId: Long): ApiStakingConfig
 
-    suspend fun getStakerStatusInfo(userAddress: String, poolAddress: String, chainId: Int): ApiStakingStatus
+    suspend fun getStakerStatusInfo(userAddress: String, poolAddress: String, chainId: Long): ApiStakingStatus
 
-    suspend fun getStakeRewardsInfo(userAddress: String, poolAddress: String, chainId: Int): ApiStakingUserRewardsInfo
+    suspend fun getStakeRewardsInfo(userAddress: String, poolAddress: String, chainId: Long): ApiStakingUserRewardsInfo
 
-    suspend fun getStakingToken(poolAddress: String, chainId: Int): ApiWalletStakingToken
+    suspend fun getStakingToken(poolAddress: String, chainId: Long): ApiWalletStakingToken
 
-    suspend fun getRewardToken(poolAddress: String, chainId: Int): ApiWalletStakingRewardsToken
+    suspend fun getRewardToken(poolAddress: String, chainId: Long): ApiWalletStakingRewardsToken
 
-    suspend fun stakeAmount(userAddress: String, amount: String, poolAddress: String, chainId: Int): ApiTransactionPerformed
+    suspend fun stakeAmount(userAddress: String, amount: String, poolAddress: String, chainId: Long): ApiTransactionPerformed
 
-    suspend fun unstakeAmount(userAddress: String, amount: String, poolAddress: String, chainId: Int): ApiTransactionPerformed
+    suspend fun unstakeAmount(userAddress: String, amount: String, poolAddress: String, chainId: Long): ApiTransactionPerformed
 
-    suspend fun claimStakingRewards(userAddress: String, poolAddress: String, chainId: Int): ApiTransactionPerformed
+    suspend fun claimStakingRewards(userAddress: String, poolAddress: String, chainId: Long): ApiTransactionPerformed
 }
