+++ zero/src/main/java/io/element/android/support/zero/network/model/request/ClaimStakingRewardsRequest.kt
@@ -12,5 +12,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ClaimStakingRewardsRequest(
     val poolAddress: String,
-    val chainId: Int
+    val chainId: Long
 )
