+++ zero/src/main/java/io/element/android/support/zero/network/model/request/StakeTransactionRequest.kt
@@ -13,5 +13,5 @@ import kotlinx.serialization.Serializable
 data class StakeTransactionRequest(
     val amount: String,
     val poolAddress: String,
-    val chainId: Int
+    val chainId: Long
 )
