+++ zero/src/main/java/io/element/android/support/zero/network/model/request/TransferWalletTokenRequest.kt
@@ -14,5 +14,5 @@ data class TransferWalletTokenRequest(
     val to: String,
     val amount: String,
     val tokenAddress: String,
-    val chainId: Int
+    val chainId: Long
 )
